Module Voqc__Qasm

val read_qasm : string -> Voqc.Main.circ * int

Read an OpenQASM file and return a circuit along with the number of qubits it uses.

val write_qasm : Voqc.Main.circ -> int -> string -> unit

Write a circuit to an OpenQASM file. The second argument is the number of qubits used in the circuit.