Module Voqc.Qasm

Functions for reading and writing files in the OpenQASM file format.

val read_qasm : string -> Main.circ * int

Read an OpenQASM file and return a circuit along with the number of qubits it uses.

val write_qasm : Main.circ -> int -> string -> unit

Write a circuit to an OpenQASM file. The second argument is the number of qubits used in the circuit.